Specified default value in action.yml.
Until now, the default value was not specified in action.yml, only in the index.js.
Therefore, it was hard to find what was in the default value.
To solve this problem, we add default value in action.yml.
And we delete the default value in index.js because the default value is specified in action.yml.